+++ tests/src/Fixtures/Basic/ImmutableDog.php
@@ -1,0 +1,26 @@
+<?php
+
+declare(strict_types=1);
+
+/*
+ * This file is part of rekalogika/mapper package.
+ *
+ * (c) Priyadi Iman Nurcahyo <https://rekalogika.dev>
+ *
+ * For the full copyright and license information, please view the LICENSE file
+ * that was distributed with this source code.
+ */
+
+namespace Rekalogika\Mapper\Tests\Fixtures\Basic;
+
+final readonly class ImmutableDog
+{
+    public function __construct(
+        private string $name,
+    ) {}
+
+    public function getName(): string
+    {
+        return $this->name;
+    }
+}