+++ tests/src/Fixtures/Basic/PersonWithImmutableDogWithoutSetter.php
@@ -1,0 +1,37 @@
+<?php
+
+declare(strict_types=1);
+
+/*
+ * This file is part of rekalogika/mapper package.
+ *
+ * (c) Priyadi Iman Nurcahyo <https://rekalogika.dev>
+ *
+ * For the full copyright and license information, please view the LICENSE file
+ * that was distributed with this source code.
+ */
+
+namespace Rekalogika\Mapper\Tests\Fixtures\Basic;
+
+class PersonWithImmutableDogWithoutSetter
+{
+    public function __construct(
+        private string $name,
+        private ImmutableDog $dog,
+    ) {}
+
+    public function getName(): string
+    {
+        return $this->name;
+    }
+
+    public function setName(string $name): void
+    {
+        $this->name = $name;
+    }
+
+    public function getDog(): ImmutableDog
+    {
+        return $this->dog;
+    }
+}